Removing default.conf otherwise Nginx fails to start
If default.conf is present nginx fails to start due to an error with the
server_names_hash length. To make this work we will have to update the
Nginx docker images to change the configuration it uses.

We will do this later, so for now we just remove the default server
configuration for Nginx.